+++ packages/json/src/rules/valueSafety.ts
@@ -90,9 +90,12 @@ export default jsonLanguage.createRule({
 		},
 	},
 	setup(context) {
-		function checkNumericLiteral(node: ts.NumericLiteral) {
-			const originalText = context.sourceFile.text.substring(
-				node.getStart(context.sourceFile),
+		function checkNumericLiteral(
+			node: ts.NumericLiteral,
+			sourceFile: ts.SourceFile,
+		) {
+			const originalText = sourceFile.text.substring(
+				node.getStart(sourceFile),
 				node.end,
 			);
 			const value = Number(originalText);
@@ -102,7 +105,7 @@ export default jsonLanguage.createRule({
 					data: { sign: value > 0 ? "" : "-" },
 					message: "infinity",
 					range: {
-						begin: node.getStart(context.sourceFile),
+						begin: node.getStart(sourceFile),
 						end: node.end,
 					},
 				});
@@ -119,7 +122,7 @@ export default jsonLanguage.createRule({
 					context.report({
 						message: "unsafeZero",
 						range: {
-							begin: node.getStart(context.sourceFile),
+							begin: node.getStart(sourceFile),
 							end: node.end,
 						},
 					});
@@ -132,7 +135,7 @@ export default jsonLanguage.createRule({
 					context.report({
 						message: "unsafeInteger",
 						range: {
-							begin: node.getStart(context.sourceFile),
+							begin: node.getStart(sourceFile),
 							end: node.end,
 						},
 					});
@@ -144,39 +147,48 @@ export default jsonLanguage.createRule({
 				context.report({
 					message: "subnormal",
 					range: {
-						begin: node.getStart(context.sourceFile),
+						begin: node.getStart(sourceFile),
 						end: node.end,
 					},
 				});
 			}
 		}
 
-		function checkStringLiteral(node: ts.StringLiteral) {
+		function checkStringLiteral(
+			node: ts.StringLiteral,
+			sourceFile: ts.SourceFile,
+		) {
 			if (hasLoneSurrogate(node.text)) {
 				context.report({
 					message: "loneSurrogate",
 					range: {
-						begin: node.getStart(context.sourceFile),
+						begin: node.getStart(sourceFile),
 						end: node.end,
 					},
 				});
 			}
 		}
 
-		function checkNode(node: ts.Node) {
+		function checkNode(node: ts.Node, sourceFile: ts.SourceFile) {
 			if (ts.isNumericLiteral(node)) {
-				checkNumericLiteral(node);
+				checkNumericLiteral(node, sourceFile);
 			} else if (ts.isStringLiteral(node)) {
-				checkStringLiteral(node);
+				checkStringLiteral(node, sourceFile);
 			} else {
-				node.forEachChild(checkNode);
+				node.forEachChild((child) => {
+					checkNode(child, sourceFile);
+				});
 			}
 		}
 
 		return {
 			visitors: {
-				ArrayLiteralExpression: checkNode,
-				ObjectLiteralExpression: checkNode,
+				ArrayLiteralExpression: (node, { sourceFile }) => {
+					checkNode(node, sourceFile);
+				},
+				ObjectLiteralExpression: (node, { sourceFile }) => {
+					checkNode(node, sourceFile);
+				},
 			},
 		};
 	},