+++ packages/json/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @flint/json
 
+## 0.15.1
+
+### Patch Changes
+
+- d99170f: fix: add missing ("phantom") dependencies to package.jsons
+- 3617e4f: chore: pass services to rule visitors
+- 3117eaf: feat: add optional async teardown() for rules
+- Updated dependencies [483ee56]
+- Updated dependencies [d99170f]
+- Updated dependencies [3617e4f]
+- Updated dependencies [f37f0d0]
+- Updated dependencies [5e23e96]
+- Updated dependencies [3117eaf]
+  - @flint.fyi/core@0.17.0
+  - @flint.fyi/rule-tester@0.15.0
+
 ## 0.15.0
 
 ### Minor Changes